use u16 only for lossless

+++ src/upsampler.rs
@@ -41,9 +41,9 @@ impl Upsampler {
         })
     }
 
-    pub fn upsample_and_interleave_row(&self, component_data: &[Vec<u16>], row: usize, output_width: usize, output: &mut [u16]) {
+    pub fn upsample_and_interleave_row(&self, component_data: &[Vec<u8>], row: usize, output_width: usize, output: &mut [u8]) {
         let component_count = component_data.len();
-        let mut line_buffer = vec![0u16; self.line_buffer_size];
+        let mut line_buffer = vec![0u8; self.line_buffer_size];
 
         debug_assert_eq!(component_count, self.components.len());
 
@@ -108,24 +108,24 @@ fn choose_upsampler(sampling_factors: (u8, u8),
 
 trait Upsample {
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     input_width: usize,
                     input_height: usize,
                     row_stride: usize,
                     row: usize,
                     output_width: usize,
-                    output: &mut [u16]);
+                    output: &mut [u8]);
 }
 
 impl Upsample for UpsamplerH1V1 {
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     _input_width: usize,
                     _input_height: usize,
                     row_stride: usize,
                     row: usize,
                     output_width: usize,
-                    output: &mut [u16]) {
+                    output: &mut [u8]) {
         let input = &input[row * row_stride ..];
 
         output[..output_width].copy_from_slice(&input[..output_width]);
@@ -134,13 +134,13 @@ impl Upsample for UpsamplerH1V1 {
 
 impl Upsample for UpsamplerH2V1 {
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     input_width: usize,
                     _input_height: usize,
                     row_stride: usize,
                     row: usize,
                     _output_width: usize,
-                    output: &mut [u16]) {
+                    output: &mut [u8]) {
         let input = &input[row * row_stride ..];
 
         if input_width == 1 {
@@ -150,28 +150,28 @@ impl Upsample for UpsamplerH2V1 {
         }
 
         output[0] = input[0];
-        output[1] = ((input[0] as u32 * 3 + input[1] as u32 + 2) >> 2) as u16;
+        output[1] = ((input[0] as u32 * 3 + input[1] as u32 + 2) >> 2) as u8;
 
         for i in 1 .. input_width - 1 {
             let sample = 3 * input[i] as u32 + 2;
-            output[i * 2]     = ((sample + input[i - 1] as u32) >> 2) as u16;
-            output[i * 2 + 1] = ((sample + input[i + 1] as u32) >> 2) as u16;
+            output[i * 2]     = ((sample + input[i - 1] as u32) >> 2) as u8;
+            output[i * 2 + 1] = ((sample + input[i + 1] as u32) >> 2) as u8;
         }
 
-        output[(input_width - 1) * 2] = ((input[input_width - 1] as u32 * 3 + input[input_width - 2] as u32 + 2) >> 2) as u16;
+        output[(input_width - 1) * 2] = ((input[input_width - 1] as u32 * 3 + input[input_width - 2] as u32 + 2) >> 2) as u8;
         output[(input_width - 1) * 2 + 1] = input[input_width - 1];
     }
 }
 
 impl Upsample for UpsamplerH1V2 {
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     _input_width: usize,
                     input_height: usize,
                     row_stride: usize,
                     row: usize,
                     output_width: usize,
-                    output: &mut [u16]) {
+                    output: &mut [u8]) {
         let row_near = row as f32 / 2.0;
         // If row_near's fractional is 0.0 we want row_far to be the previous row and if it's 0.5 we
         // want it to be the next row.
@@ -181,20 +181,20 @@ impl Upsample for UpsamplerH1V2 {
         let input_far = &input[row_far as usize * row_stride ..];
 
         for i in 0 .. output_width {
-            output[i] = ((3 * input_near[i] as u32 + input_far[i] as u32 + 2) >> 2) as u16;
+            output[i] = ((3 * input_near[i] as u32 + input_far[i] as u32 + 2) >> 2) as u8;
         }
     }
 }
 
 impl Upsample for UpsamplerH2V2 {
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     input_width: usize,
                     input_height: usize,
                     row_stride: usize,
                     row: usize,
                     _output_width: usize,
-                    output: &mut [u16]) {
+                    output: &mut [u8]) {
         let row_near = row as f32 / 2.0;
         // If row_near's fractional is 0.0 we want row_far to be the previous row and if it's 0.5 we
         // want it to be the next row.
@@ -204,37 +204,37 @@ impl Upsample for UpsamplerH2V2 {
         let input_far = &input[row_far as usize * row_stride ..];
 
         if input_width == 1 {
-            let value = ((3 * input_near[0] as u32 + input_far[0] as u32 + 2) >> 2) as u16;
+            let value = ((3 * input_near[0] as u32 + input_far[0] as u32 + 2) >> 2) as u8;
             output[0] = value;
             output[1] = value;
             return;
         }
 
         let mut t1 = 3 * input_near[0] as u32 + input_far[0] as u32;
-        output[0] = ((t1 + 2) >> 2) as u16;
+        output[0] = ((t1 + 2) >> 2) as u8;
 
         for i in 1 .. input_width {
             let t0 = t1;
             t1 = 3 * input_near[i] as u32 + input_far[i] as u32;
 
-            output[i * 2 - 1] = ((3 * t0 + t1 + 8) >> 4) as u16;
-            output[i * 2]     = ((3 * t1 + t0 + 8) >> 4) as u16;
+            output[i * 2 - 1] = ((3 * t0 + t1 + 8) >> 4) as u8;
+            output[i * 2]     = ((3 * t1 + t0 + 8) >> 4) as u8;
         }
 
-        output[input_width * 2 - 1] = ((t1 + 2) >> 2) as u16;
+        output[input_width * 2 - 1] = ((t1 + 2) >> 2) as u8;
     }
 }
 
 impl Upsample for UpsamplerGeneric {
     // Uses nearest neighbor sampling
     fn upsample_row(&self,
-                    input: &[u16],
+                    input: &[u8],
                     input_width: usize,
                     _input_height: usize,
                     row_stride: usize,
                     row: usize,
                     _output_width: usize,
-                    output: &mut [u16]) {
+                    output: &mut [u8]) {
         let mut index = 0;
         let start = (row / self.vertical_scaling_factor as usize) * row_stride;
         let input = &input[start..(start + input_width)];
